+++ components/style/lib.rs
@@ -90,10 +90,8 @@ pub mod domrefcell;
 pub mod element_state;
 pub mod error_reporting;
 pub mod font_face;
-#[cfg(feature = "gecko")] pub mod gecko_conversions;
+#[cfg(feature = "gecko")] #[allow(unsafe_code)] pub mod gecko;
 #[cfg(feature = "gecko")] #[allow(unsafe_code)] pub mod gecko_bindings;
-#[cfg(feature = "gecko")] pub mod gecko_selector_impl;
-#[cfg(feature = "gecko")] pub mod gecko_values;
 pub mod keyframes;
 pub mod logical_geometry;
 pub mod matching;
