+++ ports/geckolib/glue.rs
@@ -34,13 +34,13 @@ use style::arc_ptr_eq;
 use style::context::{LocalStyleContextCreationInfo, ReflowGoal, SharedStyleContext};
 use style::dom::{NodeInfo, TDocument, TElement, TNode};
 use style::error_reporting::StdoutErrorReporter;
-use style::gecko_selector_impl::{GeckoSelectorImpl, PseudoElement};
 use style::parallel;
 use style::parser::ParserContextExtraData;
 use style::properties::{ComputedValues, PropertyDeclarationBlock, parse_one_declaration};
 use style::selector_impl::PseudoElementCascadeType;
 use style::sequential;
 use style::stylesheets::{Origin, Stylesheet};
+use style::gecko::selector_impl::{GeckoSelectorImpl, PseudoElement};
 use style::timer::Timer;
 use traversal::RecalcStyleOnly;
 use url::Url;
